chore: DataFormat 변수 DataFormatHandler로 이동
