+++ app/src/main/java/com/github/dodobest/upbitapi/DataFormatHandler.kt
@@ -1,9 +1,32 @@
 package com.github.dodobest.upbitapi
 
 import com.github.dodobest.upbitapi.model.DataFormat
+import com.github.dodobest.upbitapi.model.MarketPlaceName
 import java.text.DecimalFormat
 
 object DataFormatHandler {
+
+    private var krwDataFormat = DataFormat(
+        changeRateFormat = "#,###.##%", priceFormat = "#,###.########"
+    )
+
+    private val btcDataFormat = DataFormat(
+        changeRateFormat = "#,###.##%", priceFormat = "#,###.########",
+    )
+
+    private val usdtDataFormat = DataFormat(
+        changeRateFormat = "#,###.##%", priceFormat = "#,###.###",
+    )
+
+    val newDataFormat = DataFormat(
+        changeRateFormat = "#,###.##%", priceFormat = "#,###.########",
+    )
+
+    val dataFormat = mapOf(
+        MarketPlaceName.KRW to krwDataFormat,
+        MarketPlaceName.BTC to btcDataFormat,
+        MarketPlaceName.USDT to usdtDataFormat,
+    )
 
     fun formatTradePrice(tradePrice: Double, dataFormat: DataFormat): String {
         return DecimalFormat(dataFormat.priceFormat).format(tradePrice)
